+++ src/components/LayoutList/index.ts
@@ -1,5 +1,4 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
 export { default as LayoutList } from './LayoutList';
-export * from './LayoutList';
+export * from './LayoutList.types';
 export { default as LayoutListItem } from './LayoutListItem';
-export * from './LayoutListItem';
